requirements.txt
python-dotenv==1.1.0
google-genai==1.15.0
google-api-python-client==2.167.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.32.0
google-cloud-monitoring==2.27.1
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2